Revert "Split veryl-translate binary from veryl"
