+++ Makefile
@@ -4,7 +4,7 @@ GIT_REVISION = $(shell git log -1 --format="%h")
 CHANNEL ?=
 VERSION = $(PKG_VERSION)$(CHANNEL) ($(GIT_REVISION) $(BUILD_DATE))
 ZIP_NAME = veryl
-BIN_NAMES = veryl veryl-ls veryl-translate
+BIN_NAMES = veryl veryl-ls
 
 export VERSION
